+++ modules/bootstrap/main.tf
@@ -9,7 +9,7 @@ terraform {
   required_providers {
     aws = {
       source  = "hashicorp/aws"
-      version = ">= 3.35"
+      version = ">= 4.0"
     }
   }
 
@@ -76,22 +76,26 @@ resource "aws_s3_bucket" "remote_state_bucket" {
   bucket = var.bucket
   tags   = var.tags
 
-  versioning {
-    enabled = true
-  }
-
-  server_side_encryption_configuration {
-    rule {
-      apply_server_side_encryption_by_default {
-        # use Amazon S3-managed keys (SSE-S3)
-        sse_algorithm = "AES256"
-      }
-    }
-  }
-
   # Terraform should never destroy this resource
   lifecycle {
     prevent_destroy = true
+  }
+}
+
+resource "aws_s3_bucket_versioning" "remote_state_bucket_versioning" {
+  bucket = aws_s3_bucket.remote_state_bucket.id
+  versioning_configuration {
+    status = "Enabled"
+  }
+}
+
+resource "aws_s3_bucket_server_side_encryption_configuration" "remote_state_bucket_encryption" {
+  bucket = aws_s3_bucket.remote_state_bucket.id
+  rule {
+    apply_server_side_encryption_by_default {
+      # use Amazon S3-managed keys (SSE-S3)
+      sse_algorithm = "AES256"
+    }
   }
 }