+++ modules/bootstrap/backend.tf
@@ -12,7 +12,6 @@ terraform {
   backend "s3" {
     region         = "us-east-2"
     dynamodb_table = "terraform"
-    encrypt        = "true"
 
     # must be unique to your AWS account; try replacing
     # uiuc-tech-services-sandbox with the friendly name of your account